basic split, need to fix indentations and imports
